requirements.txt
flask==3.0.0
tensorflow==2.15.0
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.3
Werkzeug==3.0.1
